Enable changing http request method without subclassing
To change the http request method for a QuipperService
you would've had to subclass. But this shouldn't be the
case.

So to make it easier to change the http request method
We changed it form a computed property to a stored property.

This way we could change the http request method on the
same service/api without the need for subclassing.